+++ examples/children/react16/src/index.js
@@ -57,7 +57,6 @@ window.microApp?.addGlobalDataListener(handleGlobalData);
 //   ReactDOM.unmountComponentAtNode(document.getElementById("root"));
 // })
 
-
 function mount () {
   ReactDOM.render(
     <React.StrictMode>
@@ -115,3 +114,34 @@ if (window.__MICRO_APP_ENVIRONMENT__) {
 // const dynamicStyle = document.createElement('style')
 // document.head.appendChild(dynamicStyle)
 // dynamicStyle.textContent = '.test-class { color: red } '
+
+
+// // BUG TEST: https://github.com/micro-zoe/micro-app/issues/56
+// const parser = new DOMParser()
+// const htmlString = `
+// <div>
+//   <span id='parser-id'></span>
+//   <span class='parser-class'></span>
+//   <i name='parser-name'></i>
+// </div>
+// `
+
+// const doc = parser.parseFromString(htmlString, "text/html")
+
+// console.log(
+//   'DOMParser querySelector',
+//   doc.querySelector('#parser-id'),
+//   doc.getElementById('parser-id'),
+//   doc.querySelectorAll('span'),
+//   doc.getElementsByClassName('parser-class'),
+//   doc.getElementsByTagName('span'),
+//   doc.getElementsByName('parser-name'),
+// )
+
+// setTimeout(() => {
+//   const d1 = doc.createElement('div')
+//   const d2 = doc.createElementNS('http://www.w3.org/1999/xhtml', 'svg')
+//   const d3 = doc.createDocumentFragment()
+
+//   console.log('DOMParser createElement', d1, d2, d3)
+// }, 3000)